menyederhanakan kode dan memisahkan class untuk membuat card daftar menu agar kode pada mainDaftarMenu tidak terlalu panjang
